Added auth for users

+++ posts/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required
 from .models import Post
 from .forms import PostForm
 
@@ -11,12 +12,15 @@ def home(request):
     })
 
 # Create Post
+@login_required
 def create_post(request):
     if request.method == 'POST': 
         form = PostForm(request.POST)
         
         if form.is_valid():
-            form.save()
+            post = form.save(commit=False)
+            post.author = request.user
+            post.save()
             return redirect('home')
     else:
         form = PostForm()
@@ -26,8 +30,12 @@ def create_post(request):
     })
     
 # Update Post by id
+@login_required
 def update_post(request, id): 
     post = get_object_or_404(Post, id=id)
+    
+    if post.author != request.user:
+        return redirect('home')
     
     if request.method == 'POST':
         form = PostForm(request.POST, instance=post)
@@ -51,8 +59,12 @@ def post_detail(request, id):
     })
 
 # Delete post by id
+@login_required
 def delete_post(request, id):
     post = get_object_or_404(Post, id=id)
+    
+    if post.author != request.user:
+        return redirect('home')
     
     if request.method == 'POST':
         post.delete()
